Making RNE per session
