+++ drools-core/src/main/java/org/drools/core/phreak/PhreakAsyncSendNode.java
@@ -38,17 +38,22 @@ import org.kie.internal.concurrent.ExecutorProviderFactory;
 
 public class PhreakAsyncSendNode {
 
+    private ReteEvaluator reteEvaluator;
+
+    public PhreakAsyncSendNode(ReteEvaluator reteEvaluator) {
+        this.reteEvaluator = reteEvaluator;
+    }
+
     private Executor executor() {
         return ExecutorProviderFactory.getExecutorProvider().getExecutor();
     }
 
-    public void doNode(ReteEvaluator reteEvaluator,
-                       AsyncSendNode node,
+    public void doNode(AsyncSendNode node,
                        AsyncSendMemory memory,
                        TupleSets srcLeftTuples) {
 
         if (srcLeftTuples.getInsertFirst() != null) {
-            doLeftInserts(node, memory, reteEvaluator, srcLeftTuples);
+            doLeftInserts(node, memory, srcLeftTuples);
         }
 
         srcLeftTuples.resetAll();
@@ -56,7 +61,6 @@ public class PhreakAsyncSendNode {
 
     public void doLeftInserts(AsyncSendNode node,
                               AsyncSendMemory memory,
-                              ReteEvaluator reteEvaluator,
                               TupleSets srcLeftTuples) {
 
         BetaMemory bm = memory.getBetaMemory();
@@ -71,7 +75,7 @@ public class PhreakAsyncSendNode {
 
             PropagationContext propagationContext = leftTuple.getPropagationContext();
 
-            boolean useLeftMemory = RuleNetworkEvaluator.useLeftMemory(node, leftTuple);
+            boolean useLeftMemory = RuleNetworkEvaluatorImpl.useLeftMemory(node, leftTuple);
 
             if (useLeftMemory) {
                 memory.getBetaMemory().getLeftTupleMemory().add(leftTuple);
@@ -84,8 +88,8 @@ public class PhreakAsyncSendNode {
 
             executor().execute( () -> {
                 // TODO context is not thread safe, it needs to be cloned
-                fetchAndSendResults( node, memory, reteEvaluator, context, betaConstraints, alphaConstraints, dataProvider,
-                        resultClass, finalLeftTuple, propagationContext );
+                fetchAndSendResults( node, memory, context, betaConstraints, alphaConstraints, dataProvider, resultClass,
+                        finalLeftTuple, propagationContext );
             } );
 
             leftTuple.clearStaged();
@@ -94,9 +98,9 @@ public class PhreakAsyncSendNode {
         betaConstraints.resetTuple(context);
     }
 
-    private void fetchAndSendResults(AsyncSendNode node, AsyncSendMemory memory, ReteEvaluator reteEvaluator,
-                                     Object context, BetaConstraints betaConstraints, AlphaNodeFieldConstraint[] alphaConstraints,
-                                     DataProvider dataProvider, Class<?> resultClass, TupleImpl leftTuple, PropagationContext propagationContext ) {
+    private void fetchAndSendResults(AsyncSendNode node, AsyncSendMemory memory, Object context,
+                                     BetaConstraints betaConstraints, AlphaNodeFieldConstraint[] alphaConstraints, DataProvider dataProvider,
+                                     Class<?> resultClass, TupleImpl leftTuple, PropagationContext propagationContext ) {
         for (final java.util.Iterator<?> it = dataProvider.getResults(leftTuple,
                                                                       reteEvaluator,
                                                                       memory.providerContext); it.hasNext(); ) {
@@ -108,7 +112,7 @@ public class PhreakAsyncSendNode {
             InternalFactHandle factHandle = node.createFactHandle(leftTuple, propagationContext, reteEvaluator, object);
 
             if ( isAllowed( factHandle, alphaConstraints, reteEvaluator ) ) {
-                propagate( node, reteEvaluator, factHandle, betaConstraints, context );
+                propagate( node, factHandle, betaConstraints, context );
             }
         }
     }
@@ -127,7 +131,6 @@ public class PhreakAsyncSendNode {
     }
 
     public void propagate( AsyncSendNode node,
-                           ReteEvaluator reteEvaluator,
                            InternalFactHandle factHandle,
                            BetaConstraints betaConstraints,
                            Object context ) {
