+++ drools-core/src/main/java/org/drools/core/phreak/PhreakFromNode.java
@@ -22,6 +22,7 @@ import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
+import org.drools.base.base.ValueResolver;
 import org.drools.base.rule.accessor.DataProvider;
 import org.drools.base.rule.constraint.AlphaNodeFieldConstraint;
 import org.drools.core.common.BetaConstraints;
@@ -43,8 +44,14 @@ import org.kie.api.runtime.rule.FactHandle;
 import static org.drools.core.phreak.PhreakJoinNode.updateChildLeftTuple;
 
 public class PhreakFromNode {
-    public void doNode(ReteEvaluator reteEvaluator,
-                       FromNode fromNode,
+    
+    protected ReteEvaluator reteEvaluator;
+
+    public PhreakFromNode(ReteEvaluator reteEvaluator) {
+        this.reteEvaluator = reteEvaluator;
+    }
+    
+    public void doNode(FromNode fromNode,
                        FromMemory fm,
                        LeftTupleSink sink,
                        TupleSets srcLeftTuples,
@@ -56,11 +63,11 @@ public class PhreakFromNode {
         }
 
         if (srcLeftTuples.getUpdateFirst() != null) {
-            doLeftUpdates(fromNode, fm, sink, reteEvaluator, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
+            doLeftUpdates(fromNode, fm, sink, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
         }
 
         if (srcLeftTuples.getInsertFirst() != null) {
-            doLeftInserts(fromNode, fm, sink, reteEvaluator, srcLeftTuples, trgLeftTuples);
+            doLeftInserts(fromNode, fm, sink, srcLeftTuples, trgLeftTuples);
         }
 
         srcLeftTuples.resetAll();
@@ -69,7 +76,6 @@ public class PhreakFromNode {
     public void doLeftInserts(FromNode fromNode,
                               FromMemory fm,
                               LeftTupleSink sink,
-                              ReteEvaluator reteEvaluator,
                               TupleSets srcLeftTuples,
                               TupleSets trgLeftTuples) {
 
@@ -86,7 +92,7 @@ public class PhreakFromNode {
             PropagationContext propagationContext = leftTuple.getPropagationContext();
 
             Map<Object, RightTuple> matches       = null;
-            boolean                 useLeftMemory = RuleNetworkEvaluator.useLeftMemory(fromNode, leftTuple);
+            boolean                 useLeftMemory = RuleNetworkEvaluatorImpl.useLeftMemory(fromNode, leftTuple);
 
             if (useLeftMemory) {
                 fm.getBetaMemory().getLeftTupleMemory().add(leftTuple);
@@ -124,7 +130,6 @@ public class PhreakFromNode {
     public void doLeftUpdates(FromNode fromNode,
                               FromMemory fm,
                               LeftTupleSink sink,
-                              ReteEvaluator reteEvaluator,
                               TupleSets srcLeftTuples,
                               TupleSets trgLeftTuples,
                               TupleSets stagedLeftTuples) {
@@ -205,7 +210,7 @@ public class PhreakFromNode {
                 while (childLeftTuple != null) {
                     childLeftTuple.setPropagationContext( leftTuple.getPropagationContext());
                     TupleImpl nextChild = childLeftTuple.getHandleNext();
-                    RuleNetworkEvaluator.unlinkAndDeleteChildLeftTuple( trgLeftTuples, stagedLeftTuples, childLeftTuple );
+                    RuleNetworkEvaluatorImpl.unlinkAndDeleteChildLeftTuple( trgLeftTuples, stagedLeftTuples, childLeftTuple );
                     childLeftTuple = nextChild;
                 }
             }
@@ -217,11 +222,11 @@ public class PhreakFromNode {
 
     public static boolean isAllowed( FactHandle factHandle,
                                      AlphaNodeFieldConstraint[] alphaConstraints,
-                                     ReteEvaluator reteEvaluator,
+                                     ValueResolver valueResolver,
                                      FromMemory fm ) {
         if (alphaConstraints != null) {
             for (int i = 0, length = alphaConstraints.length; i < length; i++) {
-                if (!alphaConstraints[i].isAllowed(factHandle, reteEvaluator)) {
+                if (!alphaConstraints[i].isAllowed(factHandle, valueResolver)) {
                     return false;
                 }
             }
@@ -266,7 +271,7 @@ public class PhreakFromNode {
                                             TupleImpl childLeftTuple) {
         if (childLeftTuple != null) {
             childLeftTuple.setPropagationContext( propagationContext );
-            RuleNetworkEvaluator.unlinkAndDeleteChildLeftTuple(trgLeftTuples, stagedLeftTuples, childLeftTuple);
+            RuleNetworkEvaluatorImpl.unlinkAndDeleteChildLeftTuple(trgLeftTuples, stagedLeftTuples, childLeftTuple);
         }
     }
 }
