+++ drools-core/src/main/java/org/drools/core/phreak/PhreakReactiveFromNode.java
@@ -25,15 +25,21 @@ import org.drools.core.reteoo.ReactiveFromNode;
 import org.drools.core.reteoo.ReactiveFromNode.ReactiveFromMemory;
 
 public class PhreakReactiveFromNode extends PhreakFromNode {
-    public void doNode(ReteEvaluator reteEvaluator,
-                       ReactiveFromNode fromNode,
+    
+    
+    
+    public PhreakReactiveFromNode(ReteEvaluator reteEvaluator) {
+        super(reteEvaluator);
+    }
+
+    public void doNode(ReactiveFromNode fromNode,
                        ReactiveFromMemory fm,
                        LeftTupleSink sink,
                        TupleSets srcLeftTuples,
                        TupleSets trgLeftTuples,
                        TupleSets stagedLeftTuples) {
 
-        super.doNode(reteEvaluator, fromNode, fm, sink, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
+        super.doNode(fromNode, fm, sink, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
         trgLeftTuples.addAll(fm.getStagedLeftTuples().takeAll());
     }
 }
