+++ drools-core/src/main/java/org/drools/core/phreak/PhreakTimerNode.java
@@ -53,10 +53,16 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static org.drools.core.phreak.BuildtimeSegmentUtilities.nextNodePosMask;
-import static org.drools.core.phreak.RuleNetworkEvaluator.normalizeStagedTuples;
+import static org.drools.core.phreak.RuleNetworkEvaluatorImpl.normalizeStagedTuples;
 
 public class PhreakTimerNode {
     private static final Logger log = LoggerFactory.getLogger( PhreakTimerNode.class );
+
+    private ReteEvaluator reteEvaluator;
+
+    public PhreakTimerNode(ReteEvaluator reteEvaluator) {
+        this.reteEvaluator = reteEvaluator;
+    }
 
     public void doNode(TimerNode timerNode,
                        TimerNodeMemory tm,
@@ -69,7 +75,7 @@ public class PhreakTimerNode {
                        TupleSets stagedLeftTuples) {
 
         if ( srcLeftTuples.getDeleteFirst() != null ) {
-            doLeftDeletes( timerNode, tm, pmem, sink, activationsManager, srcLeftTuples, trgLeftTuples, stagedLeftTuples );
+            doLeftDeletes( timerNode, tm, pmem, sink, srcLeftTuples, trgLeftTuples, stagedLeftTuples );
         }
 
         if ( srcLeftTuples.getUpdateFirst() != null ) {
@@ -94,10 +100,10 @@ public class PhreakTimerNode {
                               TupleSets srcLeftTuples,
                               TupleSets trgLeftTuples) {
         Timer timer = timerNode.getTimer();
-        TimerService timerService = activationsManager.getReteEvaluator().getTimerService();
+        TimerService timerService = reteEvaluator.getTimerService();
         long timestamp = timerService.getCurrentTime();
         String[] calendarNames = timerNode.getCalendarNames();
-        Calendars calendars = activationsManager.getReteEvaluator().getCalendars();
+        Calendars calendars = reteEvaluator.getCalendars();
 
         for (TupleImpl leftTuple = srcLeftTuples.getInsertFirst(); leftTuple != null; ) {
             TupleImpl next = leftTuple.getStagedNext();
@@ -121,10 +127,10 @@ public class PhreakTimerNode {
         Timer timer = timerNode.getTimer();
 
         // Variables may have changed for ExpressionIntervalTimer, so it must be rescheduled
-        TimerService timerService = activationsManager.getReteEvaluator().getTimerService();
+        TimerService timerService = reteEvaluator.getTimerService();
         long timestamp = timerService.getCurrentTime();
         String[] calendarNames = timerNode.getCalendarNames();
-        Calendars calendars = activationsManager.getReteEvaluator().getCalendars();
+        Calendars calendars = reteEvaluator.getCalendars();
 
         for ( TupleImpl leftTuple = srcLeftTuples.getUpdateFirst(); leftTuple != null; ) {
             TupleImpl next = leftTuple.getStagedNext();
@@ -145,11 +151,10 @@ public class PhreakTimerNode {
                               TimerNodeMemory tm,
                               PathMemory pmem,
                               LeftTupleSink sink,
-                              ActivationsManager activationsManager,
                               TupleSets srcLeftTuples,
                               TupleSets trgLeftTuples,
                               TupleSets stagedLeftTuples) {
-        TimerService timerService = activationsManager.getReteEvaluator().getTimerService();
+        TimerService timerService = reteEvaluator.getTimerService();
 
         TupleList leftTuples = tm.getInsertOrUpdateLeftTuples();
         TupleList deletes = tm.getDeleteLeftTuples();
@@ -194,7 +199,7 @@ public class PhreakTimerNode {
                 TupleImpl childLeftTuple = leftTuple.getFirstChild(); // only has one child
                 if ( childLeftTuple != null ) {
                     childLeftTuple.setPropagationContext( leftTuple.getPropagationContext() );
-                    RuleNetworkEvaluator.deleteChildLeftTuple( childLeftTuple, trgLeftTuples, stagedLeftTuples );
+                    RuleNetworkEvaluatorImpl.deleteChildLeftTuple( childLeftTuple, trgLeftTuples, stagedLeftTuples );
                     if ( log.isTraceEnabled() ) {
                         log.trace( "Timer Delete {}", leftTuple );
                     }
@@ -220,23 +225,22 @@ public class PhreakTimerNode {
                                    final TupleImpl leftTuple,
                                    final TupleSets trgLeftTuples,
                                    final TupleSets stagedLeftTuples) {
-        ReteEvaluator reteEvaluator = activationsManager.getReteEvaluator();
         if ( leftTuple.getPropagationContext().getReaderContext() == null ) {
-            final Trigger trigger = createTrigger( timerNode, reteEvaluator, timer, timestamp, calendarNames, calendars, leftTuple );
+            final Trigger trigger = createTrigger( timerNode, timer, timestamp, calendarNames, calendars, leftTuple );
 
             // regular propagation
-            scheduleTimer( timerNode, tm, smem, sink, reteEvaluator, timerService, timestamp, leftTuple, trgLeftTuples, stagedLeftTuples, trigger );
+            scheduleTimer( timerNode, tm, smem, sink, timerService, timestamp, leftTuple, trgLeftTuples, stagedLeftTuples, trigger );
         } else {
             // de-serializing, so we need to correlate timers before scheduling them
             Scheduler scheduler = new Scheduler() {
                 @Override
                 public void schedule( Trigger t ) {
-                    scheduleTimer( timerNode, tm, smem, sink, reteEvaluator, timerService, timestamp, leftTuple, trgLeftTuples, stagedLeftTuples, t );
+                    scheduleTimer( timerNode, tm, smem, sink, timerService, timestamp, leftTuple, trgLeftTuples, stagedLeftTuples, t );
                     evaluate( pmem, activationsManager, sink, tm, trgLeftTuples );
                 }
                 @Override
                 public Trigger getTrigger() {
-                    return createTrigger( timerNode, reteEvaluator, timer, timestamp, calendarNames, calendars, leftTuple );
+                    return createTrigger( timerNode, timer, timestamp, calendarNames, calendars, leftTuple );
                 }
             };
             TupleKey key = TupleKey.createTupleKey( leftTuple );
@@ -246,7 +250,6 @@ public class PhreakTimerNode {
     }
 
     private Trigger createTrigger(final TimerNode timerNode,
-                                  final ReteEvaluator reteEvaluator,
                                   final Timer timer,
                                   final long timestamp,
                                   final String[] calendarNames,
@@ -266,7 +269,6 @@ public class PhreakTimerNode {
                                TimerNodeMemory tm,
                                SegmentMemory smem,
                                LeftTupleSink sink,
-                               ReteEvaluator reteEvaluator,
                                TimerService timerService,
                                long timestamp,
                                TupleImpl leftTuple,
@@ -431,7 +433,7 @@ public class PhreakTimerNode {
         }
     }
 
-    private static void evaluate(PathMemory pmem,
+    private void evaluate(PathMemory pmem,
                                  ActivationsManager activationsManager,
                                  LeftTupleSink sink,
                                  TimerNodeMemory tm,
@@ -452,7 +454,7 @@ public class PhreakTimerNode {
             bit = nextNodePosMask(bit);
         }
 
-        RuleNetworkEvaluator.INSTANCE.outerEval(activationsManager, pmem.getRuleAgendaItem().getRuleExecutor(), pmem, smems,
+        reteEvaluator.getRuleNetworkEvaluator().outerEval(activationsManager, pmem.getRuleAgendaItem().getRuleExecutor(), pmem, smems,
                                                 smemIndex, bit, tm,
                                                 sink, trgLeftTuples, true);
     }
