+++ drools-metric/src/main/java/org/drools/metric/phreak/PhreakBranchNodeMetric.java
@@ -19,16 +19,20 @@
 package org.drools.metric.phreak;
 
 import org.drools.core.common.ActivationsManager;
+import org.drools.core.common.ReteEvaluator;
 import org.drools.core.common.TupleSets;
 import org.drools.core.phreak.PhreakBranchNode;
 import org.drools.core.phreak.RuleExecutor;
 import org.drools.core.reteoo.ConditionalBranchNode;
 import org.drools.core.reteoo.ConditionalBranchNode.ConditionalBranchMemory;
-import org.drools.core.reteoo.LeftTuple;
 import org.drools.core.reteoo.LeftTupleSink;
 import org.drools.metric.util.MetricLogUtils;
 
 public class PhreakBranchNodeMetric extends PhreakBranchNode {
+
+    public PhreakBranchNodeMetric(ReteEvaluator reteEvaluator) {
+        super(reteEvaluator);
+    }
 
     @Override
     public void doNode(ActivationsManager activationsManager,
