+++ drools-metric/src/main/java/org/drools/metric/phreak/PhreakExistsNodeMetric.java
@@ -28,9 +28,12 @@ import org.drools.metric.util.MetricLogUtils;
 
 public class PhreakExistsNodeMetric extends PhreakExistsNode {
 
+    public PhreakExistsNodeMetric(ReteEvaluator reteEvaluator) {
+        super(reteEvaluator);
+    }
+
     @Override
-    public void doNode(ReteEvaluator reteEvaluator,
-                       ExistsNode existsNode,
+    public void doNode(ExistsNode existsNode,
                        LeftTupleSink sink,
                        BetaMemory bm,
                        TupleSets srcLeftTuples,
@@ -40,7 +43,7 @@ public class PhreakExistsNodeMetric extends PhreakExistsNode {
         try {
             MetricLogUtils.getInstance().startMetrics(existsNode);
 
-            super.doNode(reteEvaluator, existsNode, sink, bm, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
+            super.doNode(existsNode, sink, bm, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
 
         } finally {
             MetricLogUtils.getInstance().logAndEndMetrics();
