+++ drools-metric/src/main/java/org/drools/metric/phreak/PhreakFromNodeMetric.java
@@ -29,9 +29,12 @@ import org.drools.metric.util.MetricLogUtils;
 
 public class PhreakFromNodeMetric extends PhreakFromNode {
 
+    public PhreakFromNodeMetric(ReteEvaluator reteEvaluator) {
+        super(reteEvaluator);
+    }
+
     @Override
-    public void doNode(ReteEvaluator reteEvaluator,
-                       FromNode fromNode,
+    public void doNode(FromNode fromNode,
                        FromMemory fm,
                        LeftTupleSink sink,
                        TupleSets srcLeftTuples,
@@ -41,7 +44,7 @@ public class PhreakFromNodeMetric extends PhreakFromNode {
         try {
             MetricLogUtils.getInstance().startMetrics(fromNode);
 
-            super.doNode(reteEvaluator, fromNode, fm, sink, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
+            super.doNode(fromNode, fm, sink, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
 
         } finally {
             MetricLogUtils.getInstance().logAndEndMetrics();
