+++ drools-metric/src/main/java/org/drools/metric/phreak/PhreakTimerNodeMetric.java
@@ -19,6 +19,7 @@
 package org.drools.metric.phreak;
 
 import org.drools.core.common.ActivationsManager;
+import org.drools.core.common.ReteEvaluator;
 import org.drools.core.common.TupleSets;
 import org.drools.core.phreak.PhreakTimerNode;
 import org.drools.core.reteoo.LeftTuple;
@@ -30,6 +31,10 @@ import org.drools.core.reteoo.TimerNode.TimerNodeMemory;
 import org.drools.metric.util.MetricLogUtils;
 
 public class PhreakTimerNodeMetric extends PhreakTimerNode {
+
+    public PhreakTimerNodeMetric(ReteEvaluator reteEvaluator) {
+        super(reteEvaluator);
+    }
 
     @Override
     public void doNode(TimerNode timerNode,
