+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/integrationtests/phreak/Scenario.java
@@ -177,15 +177,15 @@ public class Scenario {
         actualResultLeftTuples = new TupleSetsImpl();
         
         if ( phreakNode == PhreakJoinNode.class ) {
-            new PhreakJoinNode().doNode( wm, (JoinNode) betaNode,
-                                          sinkNode, bm, leftTuples, actualResultLeftTuples, previousResultTuples );
+            new PhreakJoinNode(wm).doNode( (JoinNode) betaNode, sinkNode,
+                                          bm, leftTuples, actualResultLeftTuples, previousResultTuples );
             
         } else if ( phreakNode == PhreakNotNode.class ) {
-            new PhreakNotNode().doNode( wm, (NotNode) betaNode,
-                                        sinkNode, bm, leftTuples, actualResultLeftTuples, previousResultTuples );            
+            new PhreakNotNode(wm).doNode( (NotNode) betaNode, sinkNode,
+                                        bm, leftTuples, actualResultLeftTuples, previousResultTuples );            
         } else if ( phreakNode == PhreakExistsNode.class ) {
-            new PhreakExistsNode().doNode( wm, (ExistsNode) betaNode,
-                                           sinkNode, bm, leftTuples, actualResultLeftTuples, previousResultTuples );            
+            new PhreakExistsNode(wm).doNode( (ExistsNode) betaNode, sinkNode,
+                                           bm, leftTuples, actualResultLeftTuples, previousResultTuples );            
         }
         
         if ( expectedResultBuilder != null ) {
